feat: add HTTP server with correction and custom dictionary
