--- cmd/server/main.go
@@ -1,93 +1,0 @@
-package main
-
-import (
-	"encoding/json"
-	"log"
-	"net/http"
-	"os"
-	"strconv"
-	"strings"
-
-	"github.com/redis/go-redis/v9"
-
-	"corrector/internal/customdict"
-)
-
-func main() {
-	redisAddr := getenv("REDIS_ADDR", "localhost:6379")
-	redisPassword := os.Getenv("REDIS_PASSWORD")
-	redisDB := getEnvInt("REDIS_DB", 0)
-
-	client := redis.NewClient(&redis.Options{
-		Addr:     redisAddr,
-		Password: redisPassword,
-		DB:       redisDB,
-	})
-
-	dict := customdict.New(client)
-
-	mux := http.NewServeMux()
-	mux.HandleFunc("/api/v1/custom-word", func(w http.ResponseWriter, r *http.Request) {
-		if r.Method != http.MethodPost {
-			http.NotFound(w, r)
-			return
-		}
-		var req struct {
-			Word string `json:"word"`
-		}
-		if err := json.NewDecoder(r.Body).Decode(&req); err != nil || req.Word == "" {
-			w.WriteHeader(http.StatusBadRequest)
-			json.NewEncoder(w).Encode(map[string]string{"error": "invalid request"})
-			return
-		}
-		if err := dict.Add(req.Word); err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			json.NewEncoder(w).Encode(map[string]string{"error": err.Error()})
-			return
-		}
-		w.WriteHeader(http.StatusCreated)
-		json.NewEncoder(w).Encode(map[string]string{"status": "ok"})
-	})
-
-	mux.HandleFunc("/api/v1/custom-word/", func(w http.ResponseWriter, r *http.Request) {
-		if r.Method != http.MethodDelete {
-			http.NotFound(w, r)
-			return
-		}
-		word := strings.TrimPrefix(r.URL.Path, "/api/v1/custom-word/")
-		if word == "" {
-			w.WriteHeader(http.StatusBadRequest)
-			json.NewEncoder(w).Encode(map[string]string{"error": "word is required"})
-			return
-		}
-		if err := dict.Remove(word); err != nil {
-			w.WriteHeader(http.StatusInternalServerError)
-			json.NewEncoder(w).Encode(map[string]string{"error": err.Error()})
-			return
-		}
-		json.NewEncoder(w).Encode(map[string]string{"status": "ok"})
-	})
-
-	addr := getenv("HTTP_ADDR", ":8080")
-	log.Printf("listening on %s", addr)
-	log.Fatal(http.ListenAndServe(addr, mux))
-}
-
-func getenv(key, def string) string {
-	v := os.Getenv(key)
-	if v == "" {
-		return def
-	}
-	return v
-}
-
-func getEnvInt(key string, def int) int {
-	v := os.Getenv(key)
-	if v == "" {
-		return def
-	}
-	if i, err := strconv.Atoi(v); err == nil {
-		return i
-	}
-	return def
-}
